Refactor task prep to happen before offers are hoarded
Previously an offer would be hoarded while the docker context was being built and uploaded
into the filesystem. This can be quite problematic with builds that have fairly large
contexts. In this commit i've refactored away the static list of tasks, and moved to
a more extensible .enqueue_build() API on the scheduler that can be called
from outside.

Closes #45 #40